requirements.txt
disnake==2.4.0
PyYAML==6.0
aiomysql==0.0.22
redis==4.1.2
progress==1.6